case GlobalRuleTemplate/ImportRuleTemplate
